test searching for a patient, change to patient search options
change to patient search options because of limitation with googles url encoder but also to remain flexible to any api changes
to avoid major breaking changes

+++ model/search.go
@@ -1,5 +1,5 @@
 package model
 
 type Search struct {
-	Score int64 `json:"score"`
+	Score float64 `json:"score"`
 }
